Additional Clang flags
• Added addition Clang warning flags `-Wall`, `-Wextra`, prompted by reading http://spin.atomicobject.com/2013/10/20/clang-compiler/ and by what I had used previously in P001.
	• Excluding warnings I don’t agree with:
		◦ `-Wno-unused-parameter` (unused parameters pop up all the time in Obj-C).
		◦ `-Wno-missing-field-initialers` (the ability to multi-initialize an entire struct with `{0}` is nice).
• Changes throughout to satisfy the warnings:
	◦ Changed struct initializer lists all over to name the members, satisfying `-Wmissing-braces`.
	◦ Changed a for-loop’s iterator type to `NSUInteger`, satisfying `-Wsign-compare`.
	◦ Added a `(size_t)` to a CFDataGetLength() call to satisfy `-Wsign-compare` (not sure why CFDataGetLength returns a signed value— NSData’s `- length` returns an `NSUInteger`).

+++ DrawInQuad/OutlineView.m
@@ -110,7 +110,7 @@
 - (NSArray *)points
 {
 	NSMutableArray *pointValueArray = [NSMutableArray arrayWithCapacity:_pointCount];
-	for (int pointI = 0; pointI < _pointCount; ++pointI)
+	for (NSUInteger pointI = 0; pointI < _pointCount; ++pointI)
 		pointValueArray[pointI] = [NSValue valueWithCGPoint:_points[pointI]];
 	
 	return pointValueArray;
